backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-chess==1.11.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
sentence-transformers==3.3.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.18
